Add URL Shortener Microservice code
This commit adds the code for the URL Shortener Microservice project. It includes the implementation of URL shortening, redirection, and validation. The main files modified/added are server.js, index.html, and style.css. The code follows the requirements outlined in the project description from the Back End Development and APIs curriculum. Additionally, the generateShortUrl() function is included to dynamically generate short URLs.

I have also tested the functionality of the API endpoints and made sure they pass the provided tests. The application can be accessed at http://localhost:3000 and successfully redirects to the original URL when visiting /api/shorturl/<short_url>.